+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/MainCarContext.kt
@@ -8,10 +8,8 @@ import com.mapbox.androidauto.car.search.CarPlaceSearchOptions
 import com.mapbox.androidauto.car.settings.CarSettingsStorage
 import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.extension.androidauto.MapboxCarMap
-import com.mapbox.navigation.base.formatter.DistanceFormatter
 import com.mapbox.navigation.core.MapboxNavigation
 import com.mapbox.navigation.core.MapboxNavigationProvider
-import com.mapbox.navigation.core.formatter.MapboxDistanceFormatter
 import com.mapbox.navigation.utils.internal.JobControl
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -31,13 +29,6 @@ class MainCarContext(
     @Deprecated(message = "This is being removed, replaced with MapboxNavigationApp")
     val mapboxNavigation: MapboxNavigation by lazy {
         MapboxNavigationProvider.retrieve()
-    }
-
-    @Deprecated(message = "This is being removed. You can construct one using MapboxNavigationApp")
-    val distanceFormatter: DistanceFormatter by lazy {
-        MapboxDistanceFormatter(
-            mapboxNavigation.navigationOptions.distanceFormatterOptions
-        )
     }
 
     val speedLimitOptions = MutableStateFlow(SpeedLimitOptions.Builder().build())