+++ libnavui-androidauto/src/test/java/com/mapbox/androidauto/MapboxCarNavigationManagerTest.kt
@@ -3,6 +3,7 @@ package com.mapbox.androidauto
 import androidx.car.app.CarContext
 import androidx.car.app.navigation.NavigationManager
 import androidx.car.app.navigation.NavigationManagerCallback
+import com.mapbox.androidauto.car.navigation.maneuver.CarManeuverMapper
 import com.mapbox.androidauto.testing.CarAppTestRule
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.base.trip.model.RouteProgress
@@ -16,15 +17,19 @@ import io.mockk.Runs
 import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
+import io.mockk.mockkObject
 import io.mockk.slot
+import io.mockk.unmockkAll
 import io.mockk.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.async
 import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.flow.collect
+import org.junit.After
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
@@ -46,6 +51,17 @@ class MapboxCarNavigationManagerTest {
     }
 
     private val sut = MapboxCarNavigationManager(carContext)
+
+    @Before
+    fun setup() {
+        mockkObject(CarManeuverMapper)
+        every { CarManeuverMapper.from(any<RouteProgress>(), any()) } returns mockk()
+    }
+
+    @After
+    fun teardown() {
+        unmockkAll()
+    }
 
     @Test
     fun `onAttached should set the NavigationManagerCallback`() {